LP III/estacionamento-victor/requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.12
django-bootstrap==0.2.4
django-bootstrap-datepicker-plus==4.0.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
typing-extensions==4.1.1
whitenoise==6.2.0
